+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Tzispa Helpers
 
 ## v0.1.1
 - Added exception rescue in Mail.send_smtp_mail
+- Added pattern Helper
+- Added web crawler helper
 
 ## v0.1.0
 - Initial release: code split from tzispa main gem